pyproject.toml
[project]
name = "ml-project"
version = "0.1.0"
description = "Probe vehicle speed patterns to traffic density/flow estimation using SUMO + ML"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "matplotlib>=3.7",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "lightgbm>=4.0",
    "tqdm>=4.65",
    "pyyaml>=6.0",
    "tensorboard>=2.13",
    "shap>=0.43",
    "optuna>=3.4",
    "scipy>=1.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.5",
]
simulation = [
    "traci>=1.18",
    "sumolib>=1.18",
]
api = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "httpx>=0.25",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
